chore(server): views integration test file tree
