+++ packages/twenty-server/test/integration/graphql/suites/view/view-field/failing-create-view-field-v2.integration-spec.ts
@@ -1,10 +1,10 @@
+import { faker } from '@faker-js/faker';
 import { createOneCoreViewField } from 'test/integration/metadata/suites/view-field/utils/create-one-core-view-field.util';
 import { extractRecordIdsAndDatesAsExpectAny } from 'test/utils/extract-record-ids-and-dates-as-expect-any';
 import {
   eachTestingContextFilter,
   type EachTestingContext,
 } from 'twenty-shared/testing';
-import { faker } from '@faker-js/faker';
 
 import { type CreateViewFieldInput } from 'src/engine/core-modules/view/dtos/inputs/create-view-field.input';
 
@@ -12,7 +12,7 @@ import {
   cleanupViewFieldTestV2,
   setupViewFieldTestV2,
   type ViewFieldTestSetup,
-} from './utils/setup-view-field-test-v2.util';
+} from '../utils/setup-view-field-test-v2.util';
 
 describe('View Field Resolver - Failing Create Operation - v2', () => {
   let testSetup: ViewFieldTestSetup;
