+++ packages/twenty-server/test/integration/graphql/suites/view/view-field/failing-delete-view-field-v2.integration-spec.ts
@@ -1,20 +1,20 @@
 import { TEST_NOT_EXISTING_VIEW_FIELD_ID } from 'test/integration/constants/test-view-ids.constants';
-import { updateOneCoreViewField } from 'test/integration/metadata/suites/view-field/utils/update-one-core-view-field.util';
+import { deleteOneCoreViewField } from 'test/integration/metadata/suites/view-field/utils/delete-one-core-view-field.util';
+import { extractRecordIdsAndDatesAsExpectAny } from 'test/utils/extract-record-ids-and-dates-as-expect-any';
 import {
   eachTestingContextFilter,
   type EachTestingContext,
 } from 'twenty-shared/testing';
-import { extractRecordIdsAndDatesAsExpectAny } from 'test/utils/extract-record-ids-and-dates-as-expect-any';
 
-import { type UpdateViewFieldInput } from 'src/engine/core-modules/view/dtos/inputs/update-view-field.input';
+import { type DeleteViewFieldInput } from 'src/engine/core-modules/view/dtos/inputs/delete-view-field.input';
 
 import {
   cleanupViewFieldTestV2,
   setupViewFieldTestV2,
   type ViewFieldTestSetup,
-} from './utils/setup-view-field-test-v2.util';
+} from '../utils/setup-view-field-test-v2.util';
 
-describe('View Field Resolver - Failing Update Operation - v2', () => {
+describe('View Field Resolver - Failing Delete Operation - v2', () => {
   let testSetup: ViewFieldTestSetup;
 
   beforeAll(async () => {
@@ -25,29 +25,26 @@ describe('View Field Resolver - Failing Update Operation - v2', () => {
     await cleanupViewFieldTestV2(testSetup.testObjectMetadataId);
   });
 
-  type UpdateViewFieldTestCase = {
-    input: UpdateViewFieldInput;
+  type DeleteViewFieldTestCase = {
+    input: DeleteViewFieldInput;
   };
 
-  const updateViewFieldTestCases: EachTestingContext<UpdateViewFieldTestCase>[] =
+  const deleteViewFieldTestCases: EachTestingContext<DeleteViewFieldTestCase>[] =
     [
       {
         title: 'non-existent view field',
         context: {
           input: {
             id: TEST_NOT_EXISTING_VIEW_FIELD_ID,
-            update: {
-              position: 1,
-            },
           },
         },
       },
     ];
 
-  it.each(eachTestingContextFilter(updateViewFieldTestCases))(
-    'should fail to update view field when $title',
+  it.each(eachTestingContextFilter(deleteViewFieldTestCases))(
+    'should fail to delete view field when $title',
     async ({ context }) => {
-      const response = await updateOneCoreViewField({
+      const response = await deleteOneCoreViewField({
         input: context.input,
         expectToFail: true,
       });
